+++ src/components/content/tools/flightInfo/FlightInfoData.vue
@@ -221,21 +221,13 @@
 </template>
 
 <script>
-import { computed, ref, watch } from 'vue';
-
-import Constants from '@/constants';
-
-import GlobalUtility from '@thzero/library_client/utility/global';
-
-import { useToolsBaseComponent } from '@/components/content/tools/toolsBase';
+import { useAppFlightInfoDataComponent } from '@/components/content/tools/flightInfo/appFlightInfoData';
+import { useAppFlightInfoDataProps } from '@/components/content/tools/flightInfo/appFlightInfoDataProps';
 
 export default {
 	name: 'flightInfoData',
 	props: {
-		modelValue: {
-			type: Object,
-			default: () => { return {}; }
-		}
+		...useAppFlightInfoDataProps
 	},
 	// extends: toolBase,
 	setup(props, context) {
@@ -261,39 +253,15 @@ export default {
 			notifyTimeout,
 			setErrorMessage,
 			setErrorTimer,
-			setNotify
-		} = useToolsBaseComponent(
-			props, 
-			context
-		);
-
-		const flightInfoInner = ref(null);
-
-		const measurementUnits = computed(() => {
-			return flightInfoInner.value?.measurementUnits ?? Constants.MeasurementUnits.english.id;
-		});
-		const measurementAcceleration = computed(() => {
-			return flightInfoInner.value ? GlobalUtility.$trans.t('measurements.' + measurementUnits.value + '.acceleration.abbr') : '';
-		});
-		const measurementAltitude = computed(() => {
-			return flightInfoInner.value ? GlobalUtility.$trans.t('measurements.' + measurementUnits.value + '.altitude.abbr') : '';
-		});
-		const measurementTime = computed(() => {
-			return flightInfoInner.value ? GlobalUtility.$trans.t('measurements.' + measurementUnits.value + '.time.abbr') : '';
-		});
-		const measurementVelocity = computed(() => {
-			return flightInfoInner.value ? GlobalUtility.$trans.t('measurements.' + measurementUnits.value + '.velocity.abbr') : '';
-		});
-
-		const valueType = (value, valueF) => {
-			return props.flightInfoInner?.dataTypes?.use ? valueF : value;
-		};
-
-		watch(() => props.modelValue,
-			(value) => {
-				flightInfoInner.value = value;
-			}
-		);
+			setNotify,
+			flightInfoInner,
+			measurementUnits,
+			measurementAcceleration,
+			measurementAltitude,
+			measurementTime,
+			measurementVelocity,
+			valueType
+		} = useAppFlightInfoDataComponent(props, context);
 
 		return {
 			correlationId,
