+++ src/components/content/tools/flightInfo/appFlightInfoDataProps.js
@@ -1,0 +1,6 @@
+export const useAppFlightInfoDataProps = {
+	modelValue: {
+		type: Object,
+		default: () => { return {}; }
+	}
+};
