+++ src/components/content/tools/flightInfo/charts/appFlightInfoChartProps.js
@@ -1,0 +1,5 @@
+export const useAppFlightInfoChartProps = {
+	chartData: {
+		type: Object
+	}
+};
